api-service/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
google-cloud-pubsub==2.33.0
pydantic==2.12.5
python-dotenv==1.0.0